refactor: move the #1449 guard into its own file and name two test helpers
CodeScene flagged permission_service.go for Low Cohesion (7.88 → 6.97) and
the new e2e test for Complex Method / Complex Conditional. All advisory,
but the cohesion one is a fair reading: the shared-config guard and the
version gate share a call site and nothing else, and permission_service.go
was already the package's largest unit. The guard moves to
permission_shared_config_gate.go, beside the test file that covers it; the
call site is unchanged. No behaviour change — same tests, still green.
